refactor: move list to separate file, add some optimizations
