+++ src/consts/index.ts
@@ -3,6 +3,10 @@ export const routes = {
   card: "/card/:movieId",
 };
 
+export const MOVIE_API_KEY = "4237669ebd35e8010beee2f55fd45546";
+
+export const pageSize = 20;
+
 export default {
   get isDev(): boolean {
     return process.env.NODE_ENV === "development";
@@ -11,5 +15,3 @@ export default {
     return process.env.NODE_ENV === "production";
   },
 };
-
-export const MOVIE_API_KEY = "4237669ebd35e8010beee2f55fd45546";
